[WinCairo][GraphicsLayerWC] missing tiles after animation https://bugs.webkit.org/show_bug.cgi?id=243721
Reviewed by Don Olmstead.

WCTileGrid::setCoverageRect turned on willRemove flag of a tile if the
tile was out of the coverage rect. But, it didn't return true to send
a tile update message. And, setCoverageRect didn't turned off the
willRemove flag for the coverage rect.

setCoverageRect should return true to send a tile upate message if a
tile will be removed. And it should turn off the willRemove flag if
the tile is in the coverage rect again.

* Source/WebKit/WebProcess/WebPage/wc/WCTileGrid.cpp:
(WebKit::WCTileGrid::ensureTile):
(WebKit::WCTileGrid::setCoverageRect):

Canonical link: https://commits.webkit.org/253274@main