[iOS] Audio session is never activated again after first deactivation https://bugs.webkit.org/show_bug.cgi?id=243797 <rdar://98062901>
Reviewed by Eric Carlson.

In the original implementation of RemoteAudioSessionProxy in 221567@main,
the return value of tryToSetActive() was saved to the ivar m_active. However
the return value is whether or not the call succeeded, not whether or not the
session is active. The fallout of this is that only the first request to
activate the audio session actually occurs; all subsequent requests fail
because we think the session is already active.

* Source/WebKit/GPUProcess/media/RemoteAudioSessionProxy.cpp:
(WebKit::RemoteAudioSessionProxy::tryToSetActive):

Canonical link: https://commits.webkit.org/253335@main