imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/select-event.html is failing https://bugs.webkit.org/show_bug.cgi?id=243791
Reviewed by Ryosuke Niwa.

Per the specification [1], we should not fire a select event if setting the selection range
doesn't result in a selection change. The issue was that setSelectionRange() could update
the selection in 2 places:
1. The first time when explicitly calling cacheSelection() explicitly
2. The second time indirectly by calling frame->selection().moveWithoutValidationTo()

Our `didChange` variable could be inaccurate because it only accounted for step 1 and not
step 2.

[1] https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#set-the-selection-range

* LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/select-event-expected.txt:
* Source/WebCore/html/HTMLTextFormControlElement.cpp:
(WebCore::HTMLTextFormControlElement::setSelectionRange):

Canonical link: https://commits.webkit.org/253342@main