AX: AccessibilityObject::visibleCharacterRange sometimes returns content from the previous viewport boundary https://bugs.webkit.org/show_bug.cgi?id=243565 rdar://96884540
Reviewed by Chris Fleizach.

This is a regression from https://bugs.webkit.org/show_bug.cgi?id=243286. That patch assumed that if the rectangle
we constructed had changed in it's `x` origin, we had moved to a new viewport boundary. This assumption is incorrect
for paragraphs that span multiple pages, since we would stop computing the start boundary after the first page break.

This patch removes that mechanism. Instead, after computing the start boundary and end boundary, we check backwards
and forwards one line to see if a rectangle built with that line is contained in the currently visible viewport.

* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::visibleCharacterRangeInternal const):
(WebCore::styleFromNode): Deleted.

Canonical link: https://commits.webkit.org/253385@main