WebKit::terminateWithReason() is declared in WebKit/Platform/spi, but defined in WebKit/Shared https://bugs.webkit.org/show_bug.cgi?id=243824 rdar://problem/98506618
Reviewed by Chris Dumez.

Move the WebKit function declaration to WebKit/Platform header and add a implementation file.
This way the SPI header only declares the SPI.
This way the WebKit function declaration and implementation are in consistent files.
This way the function implementation is logically part of WebKit/Platform, where it is also used.

This is a work to be able to compile Source/WebKit/Platform without the rest of Source/WebKit/
for unit testing purposes.

* Source/WebKit/Platform/IPC/cocoa/ConnectionCocoa.mm:
* Source/WebKit/Platform/cocoa/XPCUtilities.cpp: Copied from Source/WebKit/Platform/spi/Cocoa/ReasonSPI.h.
(WebKit::terminateWithReason):
* Source/WebKit/Platform/cocoa/XPCUtilities.h: Copied from Source/WebKit/Platform/spi/Cocoa/ReasonSPI.h.
* Source/WebKit/Platform/spi/Cocoa/ReasonSPI.h:
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/UIProcess/Launcher/cocoa/ProcessLauncherCocoa.mm:
(WebKit::terminateWithReason): Deleted.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/253461@main