[GTK][WPE] Improve the application icon passed to notification service https://bugs.webkit.org/show_bug.cgi?id=243983
Reviewed by Adrian Perez de Castro.

We currently check if the app icon is in the gtk theme to send the name.
That only works for GTK and apps having an icon in the theme. We can
also check if the application has a desktop file and get the icon from
there.

* Source/WebKit/UIProcess/Notifications/glib/NotificationService.cpp:
(WebKit::applicationIcon):
(WebKit::NotificationService::showNotification):

Canonical link: https://commits.webkit.org/253464@main