[GStreamer][MediaStream] Additional fixes for disabled video track handling https://bugs.webkit.org/show_bug.cgi?id=243776
Reviewed by Xabier Rodriguez-Calvar.

The internal audio/video appsrc elements are now named with monotonically increasing id in order to
ease debugging a bit.

As the element can receive the trackEnabledChanged notification from a non-main thread, we need to
keep track of the enabled state in a boolean flag, so that we can start pushing black frames as soon
as the notification was received.

The black frame default size is now fixed to 320x240 unless a size was previously configured, eg.
when the track was enabled. The previous sample caps can't be used because they might describe an
RTP payload and not a raw video format.

* Source/WebCore/platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:

Canonical link: https://commits.webkit.org/253467@main