PlatformMediaSessionManager::sessionWillBeginPlayback should allow resuming media elements https://bugs.webkit.org/show_bug.cgi?id=243993
In some cases, we end up trying to resume a media element after an end of interruption coming from PlatformMediaSessionManager::sessionWillBeginPlayback.
This for instance happens when returning to a WebRTC page from the YouTube application.
In that case, a first media element might be restarted (we get a notification from audio session to do it).
At that point, we might try to uninterrupt all media elements in PlatformMediaSessionManager::sessionWillBeginPlayback but with NoFlags, while the first media element was probably uninterrupted due to MayResumePlaying being given to endInterruption.
Given we already have one media element playing and the audio session is active, it is ok to try playing the other one as well.
When the audio session is not active, we uninterrupt with NoFlags, to prevent the following case:
- User is browsing the web
- User starts playing music in an app, this interrupts some media element in the web page.
- User gets back to the browser which uninterrupts silent video elements.
- Silent video elements should not trigger audible video elements to play as it will interrupt the music app.

Reviewed by NOBODY (OOPS!).

* Source/WebCore/platform/audio/PlatformMediaSessionManager.cpp:
(WebCore::PlatformMediaSessionManager::sessionWillBeginPlayback):