Regression(252534@main): LocalStorage directory is not resolved before creating sandbox extension https://bugs.webkit.org/show_bug.cgi?id=244049 rdar://98760078
Reviewed by Geoffrey Garen.

We used to resolve and create directories in WebsiteDataStore::parameters, and 252534@main removes that so we don't have
to resolve directories for the same directories each time we create parameters. The resolving and creating step was
moved to WebsiteDataStore::resolveDirectoriesIfNecessary. However, WebsiteDataStore::resolveDirectoriesIfNecessary does
not handle LocalStorage directory, and sandbox extension may fails to be created.

* Source/WebKit/UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::resolveDirectoriesIfNecessary):

Canonical link: https://commits.webkit.org/253537@main