[Alternate WebM Player] Lots of crackling noises can be heard with file when player is enabled. https://bugs.webkit.org/show_bug.cgi?id=242678 <rdar://96935878>
Reviewed by NOBODY (OOPS!).

This changes fixes to existing issues in the webm parser:
1- A WebM block can contain multiple frames, which would have
   made all of them have the same time
2- If the container using times using a different timescale than the content,
   apparent overlap due to rounding errors can appear.

For 1, we now accrue the number of frames parsed to properly determine the
actual presentation time of the frame within the block.
For 2, we check of the gap between the real end time of a frame (calculated from
the number of frames parsed and the content sampling rate) and the new block's time.
If the gap is smaller than the smallest value possible in the container's timescale,
we ignore the container's timestamp and calculate them ourselves from the actual content.

Samples exhibiting both issues are using the vorbis decoder which isn't available on
The EWS machines, making automatic testing difficult ; so tested manually.

* Source/WebCore/platform/graphics/cocoa/SourceBufferParserWebM.cpp:
(WebCore::WebMParser::resetState):
(WebCore::WebMParser::reset):
(WebCore::WebMParser::AudioTrackData::reset):
(WebCore::WebMParser::AudioTrackData::consumeFrameData):
(WebCore::SourceBufferParserWebM::setLogger):
* Source/WebCore/platform/graphics/cocoa/SourceBufferParserWebM.h:
(WebCore::WebMParser::TrackData::reset):