[libpas] Reduce TLC decommit frequency in mini-mode https://bugs.webkit.org/show_bug.cgi?id=244104
Reviewed by NOBODY (OOPS!).

Our assumption was that mini-mode VM want to shrink memory more frequently and pas_scavenger's task heaviness
is proportional to the size of entire heap. But it seems that TLC decommit is particularly costly, so doing it
frequently in mini-mode gets many CPU cycles. Let's reduce tick counts for TLC decommit in mini-mode to keep
the pace of that the same to wall time in mini mode and non mini mode.

* Source/bmalloc/bmalloc/bmalloc.cpp:
(bmalloc::api::enableMiniMode):
* Source/bmalloc/libpas/src/libpas/pas_scavenger.c:
(scavenger_thread_main):
* Source/bmalloc/libpas/src/libpas/pas_scavenger.h: