Expose text fragment matching via WebKit SPI https://bugs.webkit.org/show_bug.cgi?id=244079
Reviewed by Megan Gardner.

Add WKWebView SPI that asynchronously returns the first text fragment directive
match (if any) as a plain text string.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/dom/FragmentDirectiveParser.h:
* Source/WebCore/dom/FragmentDirectiveRangeFinder.h:
Expose text fragment directive parser and range computation entry points from WebCore.

* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _getTextFragmentMatchWithCompletionHandler:]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebViewPrivate.h:
Add SPI.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::getTextFragmentMatch):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::getTextFragmentMatch):
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:
Compute the range for the first viable text fragment, if any, flatten
it to a plain text string, and return it to the app.

* Tools/TestWebKitAPI/SourcesCocoa.txt:
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/TextFragments.mm: Added.
(TEST):
Add a test that ensures that the API works reasonably. This is
not meant to be an exhaustive test of text fragment parsing, just
a basic test of the API.

Canonical link: https://commits.webkit.org/253575@main