Simplify RenderLayerBacking::updateInternalHierarchy() https://bugs.webkit.org/show_bug.cgi?id=244102
Reviewed by Alan Bujtas.

This function had some complicated conditional logic based on which layers exit.
It's mostly building a linear hierarchy of layers, so we can simplify by building
a list of layers in parent-child order, and then just looping over the list.

`m_contentsContainmentLayer` is special because it will have two children; the
`m_backgroundLayer`, and the `m_graphicsLayer`; we manually parent the first,
and let the loop append the second.

* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateInternalHierarchy):

Canonical link: https://commits.webkit.org/253582@main