scrollIntoView() does not work on option elements https://bugs.webkit.org/show_bug.cgi?id=243517 <rdar://98475444>
Reviewed by Ryosuke Niwa.

<option> and <optgroup> are special because they don't have their own renderers;
they get painted as items by their enclosing <select> element.

Because of this `scrollIntoView()` on these elements did nothing, since `Element::scrollIntoView()`
returned early if there is no renderer.

Fix in a similar way to how `Element::getClientRects()` is implemented: add special
handling when the element is an HTMLOptionElement or HTMLOptGroupElement. `listBoxElementScrollIntoView()`
first handles the scroll in the <select>, then allows the existing call to `FrameView::scrollRectToVisible()`
to scroll the enclosing scrollers.

* LayoutTests/fast/scrolling/scrollIntoView-on-option-expected.txt: Added.
* LayoutTests/fast/scrolling/scrollIntoView-on-option.html: Added.
* LayoutTests/platform/ios-wk2/fast/scrolling/scrollIntoView-on-option-expected.txt: on iOS, the
<select> is rendered as a menu list, so the <option> has no visible representation to scroll to,
and no scrolling occurs. This matches the behavior for a menulist on macOS in all browsers.
* Source/WebCore/dom/Element.cpp:
(WebCore::listBoxElementScrollIntoView):
(WebCore::Element::scrollIntoView):

Canonical link: https://commits.webkit.org/253585@main