MediaStreamTrack.applyConstraints does not resolve when applying multiple constraints in parallel https://bugs.webkit.org/show_bug.cgi?id=244086 rdar://problem/98829928
Reviewed by Eric Carlson.

We were storing the applyConstraints promise in MediaStreamTrack.
If not resolved synchronously, there could be a chance that the MediaStreamTrack promise would get overwritten,
in case applyConstraints is called quickly on the same track.
This is now happening since we are asynchronously aplying constraints in GPUProcess.
We instead store the promise in the callback given to the backend so that we are sure to settle each promise.

Covered by added test.

* LayoutTests/fast/mediastream/applyConstraints-parallel-expected.txt: Added.
* LayoutTests/fast/mediastream/applyConstraints-parallel.html: Added.
* Source/WebCore/Modules/mediastream/MediaStreamTrack.cpp:
(WebCore::MediaStreamTrack::applyConstraints):
* Source/WebCore/Modules/mediastream/MediaStreamTrack.h:

Canonical link: https://commits.webkit.org/253588@main