AX: With isolated tree mode enabled, ignored objects sent along with published notifications cannot be queried by AX clients https://bugs.webkit.org/show_bug.cgi?id=244151 rdar://problem/98908539
Reviewed by Chris Fleizach.

Some published notifications, like AXValueChanged, include a reference
to the wrapper that changed. With isolated tree mode enabled, if the
live object associated with this wrapper is ignored, then we will never
create an isolated object for it. Then, when the AX client tries to use
the wrapper upon handling the notification, the wrapper will consider
itself detached because it has no isolated object.

With this patch, whenever we publish a notification that includes a
wrapper, we ensure that wrapper has an isolated object associated with
it. If it doesn't, we create it on the fly to ensure the AX client can
properly handle the notification.

* LayoutTests/accessibility/ignored-textfield-still-accessible-expected.txt: Added.
* LayoutTests/accessibility/ignored-textfield-still-accessible.html: Added.
* LayoutTests/platform/glib/TestExpectations: Disable new test.
* LayoutTests/platform/win/TestExpectations: Disable new test.
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::addUnconnectedNode): Added.
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.h:
(WebCore::AXIsolatedTree::treeForPageID): Added.
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
(WebCore::createIsolatedObjectForNotificationIfNeeded):
(WebCore::AXObjectCache::postTextStateChangePlatformNotification):
(WebCore::postUserInfoForChanges):
(WebCore::AXObjectCache::postTextReplacementPlatformNotification):
(WebCore::AXObjectCache::postTextReplacementPlatformNotificationForTextControl):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperBase.h:
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperBase.mm:
(-[WebAccessibilityObjectWrapperBase hasIsolatedObject]): Added.

Canonical link: https://commits.webkit.org/253629@main