Mouse automation test failing. https://bugs.webkit.org/show_bug.cgi?id=244095 rdar://98807993
Reviewed by Tim Horton.

If we do not create the event after setting the iVars, then the event is made from stale data.
In normal use, this isn't much of an issue, but when tests only make one event, then we do not
give the correct information and fail a test we would pass if performed by a human, but we
should have the events created from the most up to date information, so just move when it's created
to the correct place.

* Source/WebKit/UIProcess/ios/WKMouseGestureRecognizer.mm:
(-[WKMouseGestureRecognizer _hoverEntered:withEvent:]):

Canonical link: https://commits.webkit.org/253656@main