[WTF] Add ScopedTimer https://bugs.webkit.org/show_bug.cgi?id=244218 <rdar://98997193>
Reviewed by NOBODY (OOPS!).

This patch brings a new WTF API which allows for a quick drop-in way to
measure the runtime of pieces of code without the need to pull out a profiler.

Usage is straightforward, include the ScopedTimer header and then add
`WTFScopedTimer("any identifier");` to the scope you would like to measure.

* Source/WTF/ScopedTimer.cpp: Added.
(WTF::log):
(WTF::ScopedTimer::~ScopedTimer):
* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/Logging.cpp:
* Source/WTF/wtf/Logging.h:
* Source/WTF/wtf/ScopedTimer.h: Added.
(WTF::ScopedTimer::ScopedTimer):