Audio pitch changes on enabling / disabling the mic capture https://bugs.webkit.org/show_bug.cgi?id=243651 rdar://problem/98655838
Reviewed by Eric Carlson.

When starting microphone capture, we stop the track remote IO unit and use the VPIO unit.
Previously we were stopping the remote IO unit when asking the VPIO unit to start.
The VPIO unit may take some time to start and audio to be rendered will start to be buffered.
We are getting rid of this buffering by speeding up a bit playback, leading to higher pitch.

To prevent this buffering, we are no longer stopping the remote IO unit when starting the VPIO unit.
Instead, we wait for the first speaker callback to ask the remote IO unit to stop after hopping to main thread.

* Source/WebCore/platform/mediastream/mac/BaseAudioSharedUnit.h:
* Source/WebCore/platform/mediastream/mac/CoreAudioSharedUnit.cpp:
(WebCore::CoreAudioSharedUnit::provideSpeakerData):
(WebCore::CoreAudioSharedUnit::startInternal):
* Source/WebCore/platform/mediastream/mac/CoreAudioSharedUnit.h:

Canonical link: https://commits.webkit.org/253673@main