[Modern Media Controls] [iOS] media documents in an `<iframe>` don't autosize until the user taps to play https://bugs.webkit.org/show_bug.cgi?id=244220 <rdar://problem/98636512>
Reviewed by Jer Noble.

Media documents default to showing audio controls until the type of media is able to be determined.
Previously, this would only happen on `"error"` or `"play"`, which really ended up requiring the
user to tap on the `<video>` before the media document would recognize it as video media (type).

* Source/WebCore/Modules/modern-media-controls/media/media-document-controller.js:
(MediaDocumentController):
(MediaDocumentController.prototype.layout):
(MediaDocumentController.prototype.handleEvent): Deleted.
`AudioSupport` already handles listening for `"loadedmetadata"` and `"error"` (as well as changes to
`HTMLMediaElement.prototype.videoTracks`), so entirely rely on that for updating CSS instead of also
having separate event listeners for `"error"` and `"play"`. With this, media documents with a video
media (type) will present `<video>` controls without the user having to tap to play.

Canonical link: https://commits.webkit.org/253726@main