Non-breaking white spaces characters not checked for correctly in Scroll To Text Fragment directive search. https://bugs.webkit.org/show_bug.cgi?id=244231
Reviewed by Darin Adler and Yusuke Suzuki.

We were using isSpaceOrNewline to check if a character was a space or new line, but this
does not check for non-breaking white spaces. We really need to use the ICU function that
checks all defined white space characters in Unicode. Add a function to encapsulate that
check. It's possible that other instances of isSpaceOrNewline should actually be isUnicodeWhitespace
so we have filed a radar to check other call-sites to see if there are other places we should switch
to this new function.

Updated results for content-visibility-048 because the addition of the scroll to text fragment resources
changed the failure.

* LayoutTests/imported/w3c/resources/import-expectations.json:
* LayoutTests/imported/w3c/resources/resource-files.json:
* LayoutTests/imported/w3c/web-platform-tests/scroll-to-text-fragment/find-range-from-text-directive-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/scroll-to-text-fragment/find-range-from-text-directive-target.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/scroll-to-text-fragment/find-range-from-text-directive.html: Added.
* LayoutTests/tests-options.json:
* LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-048-expected.txt:
* Source/WTF/wtf/text/StringImpl.h:
(WTF::isUnicodeWhitespace):
* Source/WebCore/dom/FragmentDirectiveRangeFinder.cpp:
(WebCore::FragmentDirectiveRangeFinder::advanceRangeStartToNextNonWhitespace):

Canonical link: https://commits.webkit.org/253737@main