Wheel event does not stop macOS smooth keyboard scroll https://bugs.webkit.org/show_bug.cgi?id=244303 rdar://99101109
Reviewed by NOBODY (OOPS!).

No tests yet.

This is a WIP.

Currently, if you enable EventHandlerDrivenSmoothKeyboardScrolling,
scroll using the keyboard, then immediately after use the trackpad to
scroll, the webpage will scroll back to the position where the keyboard
scroll left off. This patch improves the smooth scroll setting so that a
wheel event tells the KeyboardScrollingAnimator to stop scrolling.

This WIP adds a method to KeyboardScrollingAnimator, stopImmediately
 that can be called to stop the animation without taking the time
To decelerate, which is what should happen when a wheel event
Interrupts the keyboard scroll.

This WIP also makes it so that KeyboardScrollingAnimator::beginKeyboardScrollGesture
Takes in a function to call after starting the scroll. The function that
Is passed to beginKeyboardScrollGesture by EventHandler alerts the AsyncScrollingCoordinator that a
Keyboard scroll animation has begun. Similarly, KeyboardScrollingAnimation::handleKeyUpEvent
Now takes in a function that is set as a callback to be called after the animation
Hits a velocity of 0. The two places where handleKeyUpEvent is called from
(EventHandler and ScrollController) pass in a callback function that alerts
The AsyncScrollingCoordinator that the keyboard scroll animation has stopped.

This patch also does some of the work of adding a KeyboardScrollAnimationInProgress
Property to ScrollingStateNode.

* Source/WebCore/page/EventHandler.cpp
(WebCore::EventHandler::stopKeyboardScrolling):
(WebCore::EventHandler::beginKeyboardScrollGesture):
* Source/WebCore/page/Page.cpp
(WebCore::Page::stopKeyboardScrollAnimation):
* Source/WebCore/page/scrolling/AsyncScrollingCoordinator.cpp
(WebCore::AsyncScrollingCoordinator::setKeyboardScrollAnimationInProgress): Added.
* Source/WebCore/page/scrolling/AsyncScrollingCoordinator.h
* Source/WebCore/page/scrolling/ScrollingCoordinator.h
* Source/WebCore/page/scrolling/ScrollingStateNode.h
* Source/WebCore/page/scrolling/ScrollingStateScrollingNode.cpp
(WebCore::ScrollingStateScrollingNode::ScrollingStateScrollingNode):
(WebCore:: ScrollingStateScrollingNode::setKeyboardScrollAnimationInProgress): Added.
(WebCore::ScrollingStateScrollingNode::setSynchronousScrollingReasons):
* Source/WebCore/page/scrolling/ScrollingStateScrollingNode.h
* Source/WebCore/platform/KeyboardScrollingAnimator.cpp
(WebCore::KeyboardScrollingAnimator::updateKeyboardScrollPosition):
(WebCore::KeyboardScrollingAnimator::beginKeyboardScrollGesture):
(WebCore::KeyboardScrollingAnimator::handleKeyUpEvent):
* Source/WebCore/platform/KeyboardScrollingAnimator.h
* Source/WebCore/platform/ScrollAnimator.cpp
(WebCore::ScrollAnimator::stopKeyboardScrollAnimation):
* Source/WebCore/platform/ScrollAnimator.h
* Source/WebCore/platform/ScrollableArea.cpp
(WebCore::ScrollableArea::stopKeyboardScrollAnimation):
* Source/WebCore/platform/ScrollableArea.h
* Source/WebCore/platform/ScrollingEffectsController.cpp
(WebCore::ScrollingEffectsController::stopKeyboardScrolling):
* Source/WebCore/platform/ScrollingEffectsController.h
* Source/WebKit/WebProcess/WebPage/EventDispatcher.cpp